% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_summarized.R
\name{buildSummarized}
\alias{buildSummarized}
\title{Generate summarized Read File for DE analyses}
\usage{
buildSummarized(sample.table = NULL, bam.dir = NULL, gtf = NULL,
  tx.db = NULL, mapping.mode = "Union", read.format = NULL,
  ignore.strand = FALSE, fragments = TRUE, summarized = NULL,
  output.log = NULL, filter = FALSE,
  BamFileList.yiedsize = NA_integer_, n.cores = 1, verbose = FALSE)
}
\arguments{
\item{sample.table}{A data.frame describing samples. For paired mode it must
contain 3 columns, with the names "file", "group" and "pairs". The filename
is the name in the directory supplied with the "bam.dir" parameter below.
This is not required if an existing summarized file is provided. Default=NULL}

\item{bam.dir}{Full path to location of bam files listed in the "file" column
in the sample.table provided above. This is not required if an existing
summarized file is provided. Default=NULL}

\item{gtf}{Full path to a gtf file describing the transcript coordinates to
map the RNA-seq reads to. GTF file is not required if providing a
pre-computed summarized experiment file previously generated using
buildSummarized() OR a tx.db object (below). Default = NULL}

\item{tx.db}{An R txdb object. E.g. TxDb.Dmelanogaster.UCSC.dm3.ensGene.
Default=NULL}

\item{mapping.mode}{Options are "Union", "IntersectionStrict" and
"IntersectionNotEmpty". see "mode" in ?summarizeOverlaps for explanation.
Default = "Union"}

\item{read.format}{Are the reads from single-end or paired-end data? Option
are "paired" or "single". An option must be selected. Default = NULL}

\item{ignore.strand}{Ignore strand when mapping reads? see "ignore.strand" in
?summarizeOverlaps for explanation. Default=FALSE}

\item{fragments}{When mapping.mode="paired", include reads from pairs that do
not map with their corresponding pair? see "fragments" in ?summarizeOverlaps
 for explanation. Default=TRUE}

\item{summarized}{Full path to a summarized experiment file. If
buildSummarized() has already been performed, the output summarized file,
saved in "/output.log/se.R" can be used as the input (e.g. if filtering is to
 be done). Default=NULL}

\item{output.log}{Full path to directory for output of log files and saved
summarized experiment generated.}

\item{filter}{Perform filtering of low count and missing data from the
summarized experiment file? This uses default filtering via "filterByExpr".
See ?filterByExpr for further information. Default=FALSE}

\item{BamFileList.yiedsize}{If running into memory problems. Set the number
of lines to an integer value. See "yieldSize" description in ?BamFileList for
 an explanation.}

\item{n.cores}{Number of cores to utilise for reading in Bam files. Use with
caution as can create memory issues if BamFileList.yiedsize is not
parameterised. Default=1}

\item{verbose}{Verbosity ON/OFF. Default=FALSE}
}
\value{
A summarized experiment
}
\description{
This function will create a summarized file, decribing reads
from RNA-seq experiments that overlap a set of transcript features.
Transcript features can be described as a gtf formatted table that is
imported, or using a txdb. This is designed to be straightforward and
minimised parameter for first pass batch RNA-seq analyses.
}
\examples{
## Extract summarized following example in the vignette
## load annotation tx.db for mapping reads
cat("The examples below will return a summarized experiment")

library(TxDb.Dmelanogaster.UCSC.dm3.ensGene)

## build a design table that lists the files and their grouping
file.list <- list.files(system.file("extdata", package="GenomicAlignments"),
                        recursive = TRUE,
                        pattern = "*bam$",
                        full = TRUE)

## create a sample table to be used with buildSummarized()
## must be comprised of a minimum of two columns, named "file" and "group",
## with one additional column: "pairing" if the data is paired
sample.table <- data.frame("file" = basename(file.list),
                           "group" = c("treat", "untreat"))

# extract the path to the bam directory - where to search for files listed in
## "sample.table"
bam.dir <- as.character(gsub(basename(file.list)[1], "", file.list[1]))
summarized.dm3 <- buildSummarized(sample.table = sample.table,
                                  bam.dir = bam.dir,
                                  tx.db = TxDb.Dmelanogaster.UCSC.dm3.ensGene,
                                  read.format = "paired")

}
