% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_DE.R
\name{multi.de.pairs}
\alias{multi.de.pairs}
\title{Batch - multiDE analysis of many comparisons}
\usage{
multi.de.pairs(summarised = NULL, paired = "unpaired",
  intercept = NULL, adjust.method = "BH", ruv.correct = FALSE,
  ensembl.annotate = FALSE, tx.db = NULL, plot.dir = NULL,
  output.voom = NULL, output.edger = NULL, output.deseq = NULL,
  output.combined = NULL, verbose = FALSE)
}
\arguments{
\item{summarised}{A "RangedSummarizedExperiment" object with included groups
to be analysed. For format specifications see ?read.summarised. E.g.
accessible as "summarised$group". Groups are used to automate colouring of
samples in unsupervised analyses. Default = NULL}

\item{paired}{Are the sample paired? If "paired" a paired statistical
analysis by including factors as pairs described in the "pairs" column of the
"RangedSummarizedExperiment" object in the model (accessible as
summarised$pairs). Options are "unpaired" or "paired". Default="unpaired"}

\item{intercept}{Optional ability to set the base term for fitting the model.
This is not necessary as all pairs are computed automatically. The base
term, if set, must match the name of group in "summarised$group". Default =
NULL}

\item{adjust.method}{Method used for multiple comparison adjustment of
p-values. Options are: "holm", "hochberg", "hommel", "bonferroni", "BH",
"BY", "fdr" or "none". See ?p.adjust.methods for a details description and
references. Default = "BH"}

\item{ruv.correct}{Remove Unwanted Variation (RUV)? See ?RUVr for
description. Currently only RUVr, which used the residuals is enabled and one
 factor of variation is determined. If set to TRUE and a "plot.dir" is
 provided, additional plots after RUV correction and the RUV residuals will
  be reported. Residuals are obtained through fitting a generalised linear
  model (GLM) using EdgeR. Residuals are then incorporated into the
  SummarizedExperiment object and models for DE analysis. Options = TRUE,
  FALSE. Default = FALSE.}

\item{ensembl.annotate}{If the dataset has been mapped to ensembl transcript
identifiers, obtain additional annotation of the ensembl transcripts. If set
to TRUE, a tx.db must also be provided (see below). Default = FALSE}

\item{tx.db}{An R txdb object. E.g. TxDb.Dmelanogaster.UCSC.dm3.ensGene. Must
be provided if ensembl.annotate is set to TRUE. Default = NULL}

\item{plot.dir}{Full path to directory for output of plots (pdf files). See
?diag.plots for more details. Default = NULL}

\item{output.voom}{If you wish to output the results of the Voom analysis,
provide a full path to directory for output of files. Default = NULL}

\item{output.edger}{If you wish to output the results of the EdgeR analysis,
provide a full path to directory for output of files. Default = NULL}

\item{output.deseq}{If you wish to output the results of the DEseq2 analysis,
provide a full path to directory for output of files. Default = NULL}

\item{output.combined}{consensusDE will report the results of Voom, EdgeR
and DEseq2 as a combined report. If you wish to output the results of the
COMBINED analysis, provide a full path to directory for output of files. In
addition to the combined data, it will also output the raw count and
normalised data to the same directory. Default = NULL}

\item{verbose}{Verbosity ON/OFF. Default=FALSE}
}
\value{
A list of all the comparison conducted.
## See vignette for details.
}
\description{
Given a summarised experiment generated using read.summarised()
this function will automatically perform differential expression (DE)
analysis for all possible groups using 3 different methods 1) EdgeR, 2) Voom
and 3) DEseq2. It will also output 10x diagnostic plots automatically, if the
 plotting options are selected (see ?diag.plots for more details).
}
\examples{
## Load the example data set and attach - see vignette for more details
library(airway)
data("airway")
## Name the groups of the data.
colData(airway)$group <- colData(airway)$dex
## Identify the file locations
colData(airway)$file <- rownames(colData(airway))
## Run multi.de.pairs() with RUV correction:
all.pairs.airway.ruv <- multi.de.pairs(summarised = airway,
                                       ruv.correct = TRUE,
                                       paired = "unpaired")

}
